+++ mathutils/Div.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 )
 
-func Div(num1,num2 int) (int,error){
+func Div(num1,num2 float64) (float64,error){
 	if num2 == 0{
 	 return 0,fmt.Errorf("Error : Division by zero is not allowed")
 	}
