+++ mathutils/Mul.go
@@ -1,5 +1,5 @@
 package mathutils
 
-func Mul(num1,num2 int) (int){
+func Mul(num1,num2 float64) (float64){
     return  num1 * num2
 }
